test: migrate zero-integrations-agentphone-link to bdd
