--- turbo/apps/api/src/signals/routes/__tests__/zero-integrations-agentphone-link.test.ts
@@ -1,404 +1,0 @@
-import { randomUUID } from "node:crypto";
-
-import { zeroIntegrationsAgentPhoneContract } from "@vm0/api-contracts/contracts/zero-integrations-agentphone";
-import { agentphoneVerificationSendCooldowns } from "@vm0/db/schema/agentphone-verification-send-cooldown";
-import { agentphoneUserLinks } from "@vm0/db/schema/agentphone-user-link";
-import { createStore } from "ccstate";
-import { and, eq } from "drizzle-orm";
-import { http, HttpResponse } from "msw";
-
-import { mockOptionalEnv } from "../../../lib/env";
-import { server } from "../../../mocks/server";
-import { writeDb$ } from "../../external/db";
-import { accept, setupApp, testContext } from "../../../__tests__/test-helpers";
-import {
-  createFixtureTracker,
-  createZeroRouteMocks,
-} from "./helpers/zero-route-test";
-
-interface AgentPhoneSendMessageBody {
-  readonly agent_id: string;
-  readonly to_number: string;
-  readonly body: string;
-}
-
-const context = testContext();
-const store = createStore();
-const writeDb = store.set(writeDb$);
-const mocks = createZeroRouteMocks(context);
-
-const trackPhone = createFixtureTracker(
-  async (fixture: { readonly phoneHandle: string }) => {
-    await writeDb
-      .delete(agentphoneUserLinks)
-      .where(eq(agentphoneUserLinks.phoneHandle, fixture.phoneHandle));
-  },
-);
-
-const trackVerificationSendCooldown = createFixtureTracker(
-  async (fixture: { readonly scope: string; readonly scopeKey: string }) => {
-    await writeDb
-      .delete(agentphoneVerificationSendCooldowns)
-      .where(
-        and(
-          eq(agentphoneVerificationSendCooldowns.scope, fixture.scope),
-          eq(agentphoneVerificationSendCooldowns.scopeKey, fixture.scopeKey),
-        ),
-      );
-  },
-);
-
-function uniqueId(prefix: string): string {
-  return `${prefix}_${randomUUID()}`;
-}
-
-function uniquePhone(): string {
-  const digits = randomUUID().replace(/\D/gu, "").padEnd(7, "0").slice(0, 7);
-  return `+1555${digits}`;
-}
-
-function setupAgentPhoneUser(): {
-  readonly userId: string;
-  readonly orgId: string;
-} {
-  const userId = uniqueId("user");
-  const orgId = uniqueId("org");
-  mocks.clerk.session(userId, orgId);
-  mockOptionalEnv("AGENTPHONE_AGENT_ID", "agt-test-agentphone");
-  mockOptionalEnv("AGENTPHONE_API_BASE_URL", "https://api.agentphone.to");
-  mockOptionalEnv("AGENTPHONE_API_KEY", "agentphone-test-key");
-  mockOptionalEnv("AGENTPHONE_PHONE_NUMBER", "+19039853128");
-  return { userId, orgId };
-}
-
-async function insertAgentPhoneUserLink(params: {
-  readonly phoneHandle: string;
-  readonly vm0UserId: string;
-  readonly orgId: string;
-}): Promise<void> {
-  await writeDb.insert(agentphoneUserLinks).values(params);
-  await trackPhone(Promise.resolve({ phoneHandle: params.phoneHandle }));
-}
-
-async function trackAgentPhoneVerificationCooldowns(params: {
-  readonly userId: string;
-  readonly orgId: string;
-  readonly phoneHandles: readonly string[];
-}): Promise<void> {
-  await trackVerificationSendCooldown(
-    Promise.resolve({
-      scope: "user_org",
-      scopeKey: `${params.orgId}:${params.userId}`,
-    }),
-  );
-  for (const phoneHandle of params.phoneHandles) {
-    await trackVerificationSendCooldown(
-      Promise.resolve({ scope: "phone", scopeKey: phoneHandle }),
-    );
-  }
-}
-
-async function findAgentPhoneUserLink(phoneHandle: string) {
-  const [row] = await writeDb
-    .select()
-    .from(agentphoneUserLinks)
-    .where(eq(agentphoneUserLinks.phoneHandle, phoneHandle))
-    .limit(1);
-  return row;
-}
-
-function agentPhoneSendMessage() {
-  const calls: AgentPhoneSendMessageBody[] = [];
-  const handler = http.post(
-    "https://api.agentphone.to/v1/messages",
-    async ({ request }) => {
-      const body = (await request.json()) as AgentPhoneSendMessageBody;
-      calls.push(body);
-      return HttpResponse.json({
-        id: uniqueId("apmsg"),
-        status: "sent",
-        channel: "sms",
-        from_number: "+19039853128",
-        to_number: body.to_number,
-      });
-    },
-  );
-  return { handler, calls };
-}
-
-describe("/api/integrations/agentphone/link", () => {
-  it("returns the current linked phone handle", async () => {
-    const user = setupAgentPhoneUser();
-    const phone = uniquePhone();
-    await insertAgentPhoneUserLink({
-      phoneHandle: phone,
-      vm0UserId: user.userId,
-      orgId: user.orgId,
-    });
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.getLinkStatus({
-        headers: { authorization: "Bearer clerk-session" },
-      }),
-      [200],
-    );
-
-    expect(response.body).toStrictEqual({
-      linked: true,
-      phoneHandle: phone,
-      agentPhoneNumber: "+19039853128",
-      configured: true,
-    });
-  });
-
-  it("sends a signed verification link and does not silently link the phone", async () => {
-    const user = setupAgentPhoneUser();
-    await trackAgentPhoneVerificationCooldowns({
-      ...user,
-      phoneHandles: ["+15555551212"],
-    });
-    const sendMessage = agentPhoneSendMessage();
-    server.use(sendMessage.handler);
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: "+1 (555) 555-1212" },
-      }),
-      [200],
-    );
-
-    expect(response.body).toStrictEqual({
-      phoneHandle: "+15555551212",
-      verificationSent: true,
-    });
-    expect(sendMessage.calls).toHaveLength(1);
-    expect(sendMessage.calls[0]).toStrictEqual(
-      expect.objectContaining({
-        agent_id: "agt-test-agentphone",
-        to_number: "+15555551212",
-      }),
-    );
-    expect(sendMessage.calls[0]?.body).toContain("/agentphone/connect?");
-    expect(sendMessage.calls[0]?.body).toContain(
-      "http://localhost:3002/agentphone/connect?",
-    );
-    await expect(
-      findAgentPhoneUserLink("+15555551212"),
-    ).resolves.toBeUndefined();
-  });
-
-  it("rate limits consecutive verification texts for the same user", async () => {
-    const user = setupAgentPhoneUser();
-    const firstPhone = uniquePhone();
-    const secondPhone = uniquePhone();
-    await trackAgentPhoneVerificationCooldowns({
-      ...user,
-      phoneHandles: [firstPhone, secondPhone],
-    });
-    const sendMessage = agentPhoneSendMessage();
-    server.use(sendMessage.handler);
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: firstPhone },
-      }),
-      [200],
-    );
-
-    const response = await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: secondPhone },
-      }),
-      [429],
-    );
-
-    expect(response.body).toStrictEqual({
-      error: {
-        message:
-          "Verification text was just sent. Wait a minute before trying again.",
-        code: "TOO_MANY_REQUESTS",
-      },
-    });
-    expect(sendMessage.calls).toHaveLength(1);
-  });
-
-  it("logs provider details when verification text sending is rejected", async () => {
-    const user = setupAgentPhoneUser();
-    const phone = uniquePhone();
-    await trackAgentPhoneVerificationCooldowns({
-      ...user,
-      phoneHandles: [phone],
-    });
-    server.use(
-      http.post("https://api.agentphone.to/v1/messages", () => {
-        return HttpResponse.text("provider quota exceeded", { status: 429 });
-      }),
-    );
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: phone },
-      }),
-      [503],
-    );
-
-    expect(response.body).toStrictEqual({
-      error: {
-        message: "AgentPhone verification text could not be sent",
-        code: "PROVIDER_UNAVAILABLE",
-      },
-    });
-    expect(context.mocks.axiomLogging.warn).toHaveBeenCalledWith(
-      "AgentPhone verification text provider rejected send",
-      expect.objectContaining({
-        agentphoneAgentId: "agt-test-agentphone",
-        phoneHandle: `***${phone.slice(-4)}`,
-        status: 429,
-        statusText: "Too Many Requests",
-        body: "provider quota exceeded",
-        context: "api:agentphone:link",
-      }),
-    );
-  });
-
-  it("logs fetch failures when verification text sending fails before response", async () => {
-    const user = setupAgentPhoneUser();
-    const phone = uniquePhone();
-    await trackAgentPhoneVerificationCooldowns({
-      ...user,
-      phoneHandles: [phone],
-    });
-    server.use(
-      http.post("https://api.agentphone.to/v1/messages", () => {
-        return HttpResponse.error();
-      }),
-    );
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: phone },
-      }),
-      [503],
-    );
-
-    expect(context.mocks.axiomLogging.error).toHaveBeenCalledWith(
-      "AgentPhone verification text send failed",
-      expect.objectContaining({
-        agentphoneAgentId: "agt-test-agentphone",
-        phoneHandle: `***${phone.slice(-4)}`,
-        context: "api:agentphone:link",
-        error: expect.objectContaining({
-          message: expect.any(String),
-        }),
-      }),
-    );
-  });
-
-  it("rejects empty normalized phone input", async () => {
-    setupAgentPhoneUser();
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: "abc" },
-      }),
-      [400],
-    );
-
-    expect(response.body).toStrictEqual({
-      error: expect.objectContaining({ code: "BAD_REQUEST" }),
-    });
-  });
-
-  it("rejects phone input without an explicit country code", async () => {
-    setupAgentPhoneUser();
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: "555-555-1212" },
-      }),
-      [400],
-    );
-
-    expect(response.body).toStrictEqual({
-      error: expect.objectContaining({
-        code: "BAD_REQUEST",
-        message: "Enter a phone number with country code, like +1 555 555 1212",
-      }),
-    });
-  });
-
-  it("rejects a phone handle already linked to another owner", async () => {
-    setupAgentPhoneUser();
-    const phone = uniquePhone();
-    await insertAgentPhoneUserLink({
-      phoneHandle: phone,
-      vm0UserId: uniqueId("existing-user"),
-      orgId: uniqueId("existing-org"),
-    });
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.startLink({
-        headers: { authorization: "Bearer clerk-session" },
-        body: { phoneHandle: phone },
-      }),
-      [409],
-    );
-
-    expect(response.body).toStrictEqual({
-      error: expect.objectContaining({ code: "CONFLICT" }),
-    });
-  });
-
-  it("disconnects the authenticated user's linked phone", async () => {
-    const user = setupAgentPhoneUser();
-    const phone = uniquePhone();
-    await insertAgentPhoneUserLink({
-      phoneHandle: phone,
-      vm0UserId: user.userId,
-      orgId: user.orgId,
-    });
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(
-      client.unlink({
-        headers: { authorization: "Bearer clerk-session" },
-      }),
-      [204],
-    );
-
-    expect(response.body).toBeUndefined();
-    await expect(findAgentPhoneUserLink(phone)).resolves.toBeUndefined();
-    expect(context.mocks.ably.publish).toHaveBeenCalledWith(
-      "agentphone:changed",
-      null,
-    );
-  });
-
-  it("requires authentication", async () => {
-    const client = setupApp({ context })(zeroIntegrationsAgentPhoneContract);
-
-    const response = await accept(client.getLinkStatus({ headers: {} }), [401]);
-
-    expect(response.body).toStrictEqual({
-      error: {
-        message: "Not authenticated",
-        code: "UNAUTHORIZED",
-      },
-    });
-  });
-});
